split posterior utils
